get_xxx_or_noneの中身を修正
